+++ TwoValueCommand.java
@@ -1,3 +1,3 @@
 interface TwoValueCommand {
-  float calculate(float a, float b);
+  double calculate(double a, double b);
 }
